Add task for pushing to Launchpad PPA
A package is published to Launchpad PPA by uploading a source `.changes`
file with `dput`. This means that we cannot build the DEB file in a
Concourse pipeline and upload it.

This commit introduces a pattern for taking an tar-archive of already
compiled Greenplum Database and structuring it in such a way that
`debuild` sees it as a source package that only contains binary blobs.
The high-level idea is the make source archive that consists of desired
file structure and building is simply copying from the arhive to the
filesystem.

With this pattern, we can build a DEB file locally to verify
functionality and then create a signed source package. This source
package can then be uploaded to Launchpad PPA for distribution.

Co-authored-by: Mark Sliva <msliva@pivotal.io>
Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>
Co-authored-by: Bradford D. Boyle <bboyle@pivotal.io>