Use non-standard upstream version for 6.0.0 GA release
Beta releases of GPDB were published with upstream versions that looked
like `6.0.0-beta.x` (i.e., semver versions). Debian version sorting is
different from semi-semantic and 6.0.0 is considered an older release
than 6.0.0-beta.7. This caused Launchpad to reject the upload for GPDB
6.0.0.

We have previously fixed the major difference between semi-semantic and
Debian version sorting by converting `-` in the upstream version to `~`.
This preserves the order of versions; for example

6.0.0-beta.7 < 6.0.0 (semi-semantic)
6.0.0~beta.7 < 6.0.0 (Debian)

Once we increase 6.0.0 to +6.0.1 the out-of-order package versions will
stop being an issue.

Rather than delete the existing beta packages in ppa:greenplum/db and
potentially break existing open-source users ability to upgrade to
6.0.0, this commit uploads with the upstream version set to 6.0.0-ga. In
Debian sorting, this is considered a newer release than 6.0.0-beta.7 and
an older release than either 6.0.1 or 6.1.0. After the GA package is
successfully published on Launchpad, this commit will be reverted.