Move non-secrets out of gp-continuous-integration for greenplum-database-release
Originally, the non-secrets variables are stored under gp-continuous-integration.
This makes changing the pipeline complicated, since we have to touch two repos
to update the pipeline.

Now, we moved the non-secrets variables into the same repo as the pipeline. Hence,
we can reduce the unnecessary coupling between greenplum-database-release and
gp-continuous-integration repos.

Co-authored-by: Ning Wu <nwu@pivotal.io>
Co-authored-by: Shaoqi Bai <sbai@pivotal.io>
Co-authored-by: Xin Zhang <xzhang@pivotal.io>
Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>