Fix build failure due after ORCA merged for OSS
Before ORCA was merged in to GPDB 6X_STABLE, it was built from an upstream
tarball and installed in to /usr/local. When archiving the GBPD build, we
copied the ORCA libraries from this prefix in to $GPHOME.  After the merge, the
ORCA libs are built by the GPDB build system and do not need to be explicitly
copied.

We had attempted to "feature flag" where to compile ORCA from *before* the
merge, but missed applying this to the vendoring function. This commit removes
INTEGRATE_GPORCA in related scripts.

[#172441364]

Co-authored-by: Bradford D. Boyle <bboyle@pivotal.io>
Co-authored-by: Xin Zhang <xzhang@pivotal.io>