GP 5: Support local RPM builds with BIN_GPDB_TARGZ
The code for building a GP 5 RPM locally assumed that you had a
tar archive named "bin_gpdb/bin_gpdb.tar.gz" in the current working
directory (i.e., in "greenplum-database-release"). This commit uses the
value of the environment variable BIN_GPDB_TARGZ to determine which tar
archive to build in to an RPM.

If the environment variable is not set or does not point to a regular
file, the code prints an error message and exits.

[#173785658]

Co-authored-by: Brent Doil <bdoil@vmware.com>
Co-authored-by: Bradford D. Boyle <bradfordb@vmware.com>