Don't compile python3.9 scripts when packaging rpm. (#215)
In 6.22 release, we introduce python39 to Greenplum. When packing the
rpm package, we compiling python scripts to .pyc bytecode files.
However, the compilation steps is using a python2 interpreter which is
unable to compile python3.9 scripts. On the other hand, python3.9
scripts have been compiled to bytecode files, we don't need to compile
them again.

Co-authored-by: Hao Zhang <hzhang2@vmware.com>
Co-authored-by: Xing Guo <higuoxing@gmail.com>